Initial commit for myFirstProject
In this initial commit, I'm excited to present myFirstProject, the first step of my coding journey as a budding programmer at the age of 16. This project marks the beginning of my pursuit to become a skilled software engineer.

myFirstProject showcases a simple yet fundamental program that I've written to learn the ropes of coding. As a beginner, I'm thrilled to embark on this adventure of exploration and learning. Through this project, I aim to lay a strong foundation for my future endeavors and contribute to the vast world of software development.

Stay tuned for more updates as I dive deeper into the world of coding, pushing my boundaries, and transforming my ideas into reality. Your support and feedback are greatly appreciated as I grow and evolve as a programmer. Here's to the exciting journey ahead! 🚀👩‍💻